Adjust tipping and contribution code to allow for non-integral tips.
Resolves #6481: Contribution code should use doubles for currency, not integers

+++ browser/extensions/api/brave_rewards_api.cc
@@ -650,7 +650,7 @@ void BraveRewardsGetRecurringTipsFunction::OnGetRecurringTips(
     for (auto const& item : *list) {
       auto tip = std::make_unique<base::DictionaryValue>();
       tip->SetString("publisherKey", item.id);
-      tip->SetInteger("amount", item.weight);
+      tip->SetDouble("amount", item.weight);
       recurringTips->Append(std::move(tip));
     }
   }
